scratch/requirements_old.txt
Flask>=3.0,<4.0
gunicorn
psycopg2-binary
flask-cors
